major work, working in Fortran - waiting for debug

+++ src/prototype/Fortran/boundary.f90
@@ -4,10 +4,10 @@
 subroutine boundary(mbar,fbar,mleft,mright,fup,fdown,N,Q)
     implicit none
     integer, intent(in) :: N,Q;
-    real, dimension(N+2,Q+1), intent(in) :: mbar;
-    real, dimension(N+1,Q+2), intent(in) :: fbar;
-    real, dimension(Q+1), intent(out) :: mleft,mright;
-    real, dimension(N+1), intent(out) :: fup,fdown;
+    double precision, dimension(N+2,Q+1), intent(in) :: mbar;
+    double precision, dimension(N+1,Q+2), intent(in) :: fbar;
+    double precision, dimension(Q+1), intent(out) :: mleft,mright;
+    double precision, dimension(N+1), intent(out) :: fup,fdown;
 
     mleft = mbar(1,:);
     mright = mbar(N+2,:);
