+++ src/prototype/Fortran/cost.f90
@@ -3,8 +3,8 @@
 subroutine cost(R,m,f,N,Q)
     implicit none
     integer, intent(in) :: N,Q;
-    real, dimension(N+1,Q+1), intent(in) :: m,f; 
-    real, intent(out) :: R;
+    double precision, dimension(N+1,Q+1), intent(in) :: m,f; 
+    double precision, intent(out) :: R;
 
     R = sum(m**2/(1.0*f));
 end subroutine
