+++ src/prototype/Fortran/proxJ.f90
@@ -1,11 +1,11 @@
 subroutine proxJ(Pm,Pf,m,f,g,N,Q)
     implicit none
     integer, intent(in) :: N,Q;
-    real, intent(in) :: g;
-    real, dimension(N+1,Q+1), intent(in) :: m,f; 
-    real, dimension(N+1,Q+1), intent(out) :: Pm,Pf; 
-    real, dimension(N+1,Q+1) :: x0,x1; 
-    real, dimension(N+1,Q+1) :: P,dP;
+    double precision, intent(in) :: g;
+    double precision, dimension(N+1,Q+1), intent(in) :: m,f; 
+    double precision, dimension(N+1,Q+1), intent(out) :: Pm,Pf; 
+    double precision, dimension(N+1,Q+1) :: x0,x1; 
+    double precision, dimension(N+1,Q+1) :: P,dP;
     integer :: k;
     
     x0 = 1000;
@@ -25,16 +25,16 @@ end subroutine
 
 subroutine polynome(P,x,m,f,g,N,Q)
     integer, intent(in) :: N,Q;
-    real, dimension(N+1,Q+1), intent(in) :: x,m,f;
-    real, dimension(N+1,Q+1), intent(out) :: P;
-    real :: g; 
+    double precision, dimension(N+1,Q+1), intent(in) :: x,m,f;
+    double precision, dimension(N+1,Q+1), intent(out) :: P;
+    double precision :: g; 
     P = -g*m**2 + (x-f)*((x+2*g)**2);
 end subroutine
 
 subroutine polynome_derivative(dP,x,f,g,N,Q)
     integer, intent(in) :: N,Q;
-    real, dimension(N+1,Q+1), intent(in) :: x,f;
-    real, dimension(N+1,Q+1), intent(out) :: dP;
-    real :: g;
+    double precision, dimension(N+1,Q+1), intent(in) :: x,f;
+    double precision, dimension(N+1,Q+1), intent(out) :: dP;
+    double precision :: g;
     dP = 2*x*(x-f) + (x+2*g)**2;
 end subroutine
